requirements.txt
PyQt6==6.8.1
pyautogui==0.9.54
requests==2.32.3
anthropic==0.49.0
python-dotenv==1.1.0
pillow==11.1.0
numpy==2.2.4
qtawesome==1.4.0
SpeechRecognition==3.14.2
pyttsx3==2.98
keyboard==0.13.5
pyaudio==0.2.14
tiktoken==0.6.0
psutil==5.9.8  # For system monitoring and process management
xlib==0.21  # For X11 fallback support
